Use the button compass on mobile
Extract the desktop button compass into a shared commonMain CompassButtons
parameterized by CompassButtonColors. Desktop's BUTTONS style delegates to
it (gameChrome colors); the mobile GameView now always uses it (Material
colors) in place of the skin compass rose.
